Bug fix: update normalised_to, not just `to` after letter sanitise
When a precompiled letter is sent to us, we set the `to` field as
'Provided as PDF' in
https://github.com/alphagov/notifications-api/blob/1c1023a877d7c7da6ab4524a9fd21f930b7fbe1f/app/v2/notifications/post_notifications.py#L100-L104

This then also sets `normalised_to` as `providedaspdf`.

However, when template preview sanitises the letter, pulls out the
address and gives it to the API, we were only setting `to` to be
the new address and had forgotten to also amend `normalised_to` to
be the normalised version. This meant that for all these letters
we accidentally left `normalised_to` as `providedaspdf`. The impact
of this was that we can not then search for these letters in the
admin user interface as they rely on the `normalised_to` field
containing the recipient address.

This commit fixes that bug by also setting the `normalised_to`
field